Moved HashStats to a separate module
